requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
motor>=3.1.0
pymongo>=4.4.0
flask>=2.2,<3
pydantic>=1.10.0
python-dotenv>=0.21.0
pytest>=7.0.0
